tg_phone_bot/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohappyeyeballs==2.4.6
aiohttp==3.8.6
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anticaptchaofficial==1.0.62
anyio==4.8.0
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
cachetools==4.2.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
Django==5.1
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kombu==5.4.2
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
playwright==1.50.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.1.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
TelegramBotAPI==0.3.2
tornado==6.4.2
typing_extensions==4.7.1
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
